ue4's build function now always builds Editor binaries
Justification:
- Packaging a project builds Editor and non-Editor binaries, regardless of configuration.
- Running AutomationTests only works when Editor binaries have been built (but doesn't need non-Editor binaries).

If someone were to only want to run tests, they'd have to build Editor binaries. However, they would not need project binaries in this case.
Therefore, it is valid to say that Editor binaries are always needed, regardless of specified configuration. Project binaries, however, are not.

By having the build function build Editor binaries and potentially project binaries, the time spent in the package function is reduced, which leads to more accurate times for pipeline steps if build- and package steps are separated.
Additionally, this also results in pipeline failures being placed in a more appropriate step (i.e. packaging did not go wrong, but building the binaries did).